Refactor codes, remove unused test codes
