+++ fastmsa/core/_errors.py
