+++ fastmsa/logging.py
@@ -1,1 +1,1 @@
-from .core.logging import get_logger  # noqa
+from .core._logging import get_logger  # noqa
